Add support for precomiling native code
Adds a compile provider to the atomvm namespace for setting the target architecture. The default
target arch is `emu`, which does not do any jit precompiling. The `compile` task configuration arch
will be used by the packbeam task, if present, in rebar.config. The --arch (-t) option may be used
directly with the packbeam task to override any compile task configuration in rebar.config.

Signed-off-by: Winford <winford@object.stream>